+++ modules/fireback/json-tools.go
@@ -1,35 +1,15 @@
-/*
-Contains a general csv operation code which can be used on different type of structs in golang
-*/
 package fireback
 
 import (
+	"encoding/json"
 	"os"
-
-	"github.com/gocarina/gocsv"
 )
 
-type ProgressUpdate struct {
-	ItemsProcessed int
-	Message        string
-	Complete       bool
-	Error          error
-}
-
-// CSV2Exporter exports data from a channel to a CSV format using the provided io.Writer.
-func CSV2ExporterWriter[T any](source chan []*T, fp string) (chan ProgressUpdate, error) {
+func JsonExporterWriter[T any](source chan []*T, fp string) (chan ProgressUpdate, error) {
 	progress := make(chan ProgressUpdate) // Channel to send progress updates
-	value, err := gocsv.MarshalBytes([]T{})
-	if err != nil {
-		return nil, err
-	}
 
 	writer, err := os.Create(fp)
 	if err != nil {
-		return nil, err
-	}
-
-	if _, err := writer.Write(value); err != nil {
 		return nil, err
 	}
 
@@ -46,9 +26,8 @@ func CSV2ExporterWriter[T any](source chan []*T, fp string) (chan ProgressUpdate
 
 			if len(batch) >= batchSize {
 
-				if res, err := gocsv.MarshalStringWithoutHeaders(batch); err == nil {
+				if res, err := json.MarshalIndent(batch, "", "  "); err == nil {
 					if _, err := writer.Write([]byte(res)); err != nil {
-						// return err
 						progress <- ProgressUpdate{
 							ItemsProcessed: 0,
 							Error:          err,
@@ -65,7 +44,7 @@ func CSV2ExporterWriter[T any](source chan []*T, fp string) (chan ProgressUpdate
 		}
 
 		if len(batch) > 0 {
-			if res, err := gocsv.MarshalStringWithoutHeaders(batch); err == nil {
+			if res, err := json.MarshalIndent(batch, "", "  "); err == nil {
 				if _, err := writer.Write([]byte(res)); err != nil {
 					progress <- ProgressUpdate{
 						ItemsProcessed: 0,
